requirements.txt
httpx
aiofiles
ruamel.yaml
beautifulsoup4
fastapi[all]